+++ cynic-querygen/src/query_parsing/mod.rs
@@ -28,14 +28,17 @@ pub fn parse_query_document<'text>(
     enums.sort_by_key(|e| e.name);
     scalars.sort_by_key(|s| s.0);
 
-    let mut query_namer = Namer::new();
+    let mut namers = Namers::new();
 
     let arg_struct_details = arguments::build_argument_structs(&normalised);
 
     for operation in &normalised.operations {
         let operation_name = operation.name.unwrap_or("UnnamedQuery");
 
-        query_namer.force_name(&operation.root, operation_name);
+        namers
+            .selection_sets
+            .force_name(&operation.root, operation_name);
+
         arg_struct_details.force_name_argument_struct_for(
             &operation.root,
             format!("{}Arguments", operation_name),
@@ -44,7 +47,13 @@ pub fn parse_query_document<'text>(
 
     let query_fragments = sorting::topological_sort(normalised.selection_sets.iter().cloned())
         .into_iter()
-        .map(|selection| make_query_fragment(selection, &mut query_namer, &arg_struct_details))
+        .map(|selection| make_query_fragment(selection, &mut namers, &arg_struct_details))
+        .collect::<Vec<_>>();
+
+    let inline_fragments = normalised
+        .inline_fragments
+        .into_iter()
+        .map(|fragment| make_inline_fragments(fragment, &mut namers, &arg_struct_details))
         .collect::<Vec<_>>();
 
     let input_objects = sorting::topological_sort(input_objects.into_iter())
@@ -54,6 +63,7 @@ pub fn parse_query_document<'text>(
 
     Ok(Output {
         query_fragments,
+        inline_fragments,
         input_objects,
         enums,
         scalars,
@@ -63,8 +73,8 @@ pub fn parse_query_document<'text>(
 
 fn make_query_fragment<'text>(
     selection: Rc<normalisation::SelectionSet<'text, 'text>>,
-    namer: &mut Namer<Rc<normalisation::SelectionSet<'text, 'text>>>,
-    argument_struct_details: &ArgumentStructDetails<'text, 'text, '_>,
+    namers: &mut Namers<'text>,
+    argument_struct_details: &ArgumentStructDetails<'text, 'text>,
 ) -> crate::output::QueryFragment<'text, 'text> {
     use crate::output::query_fragment::{
         FieldArgument, OutputField, QueryFragment, RustOutputFieldType,
@@ -75,38 +85,58 @@ fn make_query_fragment<'text>(
         fields: selection
             .selections
             .iter()
-            .map(|selection| match selection {
-                Selection::Field(field) => {
-                    let schema_field = &field.schema_field;
+            .map(|selection| {
+                let Selection::Field(field) = selection;
+                let schema_field = &field.schema_field;
 
-                    let inner_type_name = match &field.field {
-                        Field::Leaf => schema_field.value_type.inner_name().to_string(),
-                        Field::Composite(ss) => namer.name_subject(ss),
-                    };
-
-                    OutputField {
-                        name: field.alias.unwrap_or(schema_field.name),
-                        rename: field.alias.map(|_| schema_field.name),
-                        field_type: RustOutputFieldType::from_schema_type(
-                            &schema_field.value_type,
-                            inner_type_name,
-                        ),
-                        arguments: field
-                            .arguments
-                            .iter()
-                            .map(|(name, value)| -> Result<FieldArgument, Error> {
-                                Ok(FieldArgument::new(name, value.clone()))
-                            })
-                            .collect::<Result<Vec<_>, _>>()
-                            .unwrap(),
+                let inner_type_name = match &field.field {
+                    Field::Leaf => schema_field.value_type.inner_name().to_string(),
+                    Field::Composite(ss) => namers.selection_sets.name_subject(ss),
+                    Field::InlineFragments(fragments) => {
+                        namers.inline_fragments.name_subject(fragments)
                     }
+                };
+
+                OutputField {
+                    name: field.alias.unwrap_or(schema_field.name),
+                    rename: field.alias.map(|_| schema_field.name),
+                    field_type: RustOutputFieldType::from_schema_type(
+                        &schema_field.value_type,
+                        inner_type_name,
+                    ),
+                    arguments: field
+                        .arguments
+                        .iter()
+                        .map(|(name, value)| -> Result<FieldArgument, Error> {
+                            Ok(FieldArgument::new(name, value.clone()))
+                        })
+                        .collect::<Result<Vec<_>, _>>()
+                        .unwrap(),
                 }
             })
             .collect(),
         argument_struct_name: argument_struct_details.argument_name_for_selection(&selection),
 
-        name: namer.name_subject(&selection),
+        name: namers.selection_sets.name_subject(&selection),
         target_type: selection.target_type.name().to_string(),
+    }
+}
+
+fn make_inline_fragments<'text>(
+    inline_fragments: Rc<normalisation::InlineFragments<'text, 'text>>,
+    namers: &mut Namers<'text>,
+    _argument_struct_details: &ArgumentStructDetails<'text, 'text>,
+) -> crate::output::InlineFragments {
+    crate::output::InlineFragments {
+        inner_type_names: inline_fragments
+            .inner_selections
+            .iter()
+            .map(|s| namers.selection_sets.name_subject(s))
+            .collect(),
+        target_type: inline_fragments.abstract_type.name().to_string(),
+        // TODO: Deal with argument structs
+        argument_struct_name: None,
+        name: namers.inline_fragments.name_subject(&inline_fragments),
     }
 }
 
@@ -114,5 +144,20 @@ fn make_input_object(input: Rc<inputs::InputObject>) -> crate::output::InputObje
     crate::output::InputObject {
         name: input.schema_type.name.to_string(),
         fields: input.fields.clone(),
+    }
+}
+
+#[derive(Debug)]
+struct Namers<'text> {
+    selection_sets: Namer<Rc<normalisation::SelectionSet<'text, 'text>>>,
+    inline_fragments: Namer<Rc<normalisation::InlineFragments<'text, 'text>>>,
+}
+
+impl<'text> Namers<'text> {
+    pub fn new() -> Self {
+        Namers {
+            selection_sets: Namer::new(),
+            inline_fragments: Namer::new(),
+        }
     }
 }